+++ src/app/(route)/goal/editor/components/GoalEditor/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useTempData } from "./useTempData/useTempData";
